+++ analytics/models.py
@@ -12,8 +12,10 @@ from follow.models import FollowOrganizationList
 from organization.models import Organization
 import pytz
 import wevote_functions.admin
-from wevote_functions.functions import convert_date_as_integer_to_date, convert_date_to_date_as_integer, \
-    convert_to_int, positive_value_exists
+# from wevote_functions.functions import convert_date_as_integer_to_date, convert_date_to_date_as_integer, \
+#     convert_to_int, positive_value_exists
+from wevote_functions.functions import convert_to_int, positive_value_exists
+from wevote_functions.functions_time import convert_date_as_integer_to_date, convert_date_to_date_as_integer
 from wevote_settings.models import WeVoteSetting, WeVoteSettingsManager
 
 ACTION_VOTER_GUIDE_VISIT = 1
