+++ candidate/views_admin.py
@@ -44,12 +44,18 @@ from twitter.models import TwitterLinkPossibility, TwitterUserManager
 from voter.models import fetch_voter_from_voter_device_link, VoterDeviceLinkManager, VoterManager, voter_has_authority
 from volunteer_task.models import VOLUNTEER_ACTION_POLITICIAN_DEDUPLICATION, VolunteerTaskManager
 from voter_guide.models import VoterGuide
+# from wevote_functions.functions import convert_to_int, \
+#     convert_we_vote_date_string_to_date_as_integer, \
+#     extract_instagram_handle_from_text_string, \
+#     extract_twitter_handle_from_text_string, get_voter_api_device_id, get_voter_device_id, list_intersection, \
+#     positive_value_exists, STATE_CODE_MAP, display_full_name_with_correct_capitalization, \
+#     extract_state_from_ocd_division_id
 from wevote_functions.functions import convert_to_int, \
-    convert_we_vote_date_string_to_date_as_integer, \
-    extract_instagram_handle_from_text_string, \
-    extract_twitter_handle_from_text_string, get_voter_api_device_id, get_voter_device_id, list_intersection, \
+    extract_instagram_handle_from_text_string, extract_twitter_handle_from_text_string, \
+    get_voter_api_device_id, get_voter_device_id, list_intersection, \
     positive_value_exists, STATE_CODE_MAP, display_full_name_with_correct_capitalization, \
     extract_state_from_ocd_division_id
+from wevote_functions.functions_time import convert_we_vote_date_string_to_date_as_integer
 from wevote_settings.constants import ELECTION_YEARS_AVAILABLE
 from wevote_settings.models import RemoteRequestHistory, \
     RETRIEVE_POSSIBLE_GOOGLE_LINKS, RETRIEVE_POSSIBLE_TWITTER_HANDLES
