+++ office/views_admin.py
@@ -29,11 +29,16 @@ from position.models import OPPOSE, PositionListManager, SUPPORT
 from volunteer_task.models import VOLUNTEER_ACTION_CANDIDATE_CREATED, VolunteerTaskManager
 from voter.models import fetch_voter_from_voter_device_link, voter_has_authority
 import wevote_functions.admin
-from wevote_functions.functions import convert_to_int, convert_we_vote_date_string_to_date_as_integer, \
-    get_voter_api_device_id, \
+# from wevote_functions.functions import convert_to_int, convert_we_vote_date_string_to_date_as_integer, \
+#     get_voter_api_device_id, \
+#     extract_first_name_from_full_name, extract_middle_name_from_full_name, extract_last_name_from_full_name, \
+#     is_candidate_we_vote_id, is_politician_we_vote_id, positive_value_exists, \
+#     STATE_CODE_MAP
+from wevote_functions.functions import convert_to_int, get_voter_api_device_id, \
     extract_first_name_from_full_name, extract_middle_name_from_full_name, extract_last_name_from_full_name, \
     is_candidate_we_vote_id, is_politician_we_vote_id, positive_value_exists, \
     STATE_CODE_MAP
+from wevote_functions.functions_time import convert_we_vote_date_string_to_date_as_integer
 from django.http import HttpResponse
 import json
